pyproject.toml
[project]
name = "gethired"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.37.0",
    "celery>=5.6.3",
    "django>=6.0.3",
    "django-celery-beat>=2.9.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-genai>=0.1.0",
    "pgvector>=0.4.2",
    "pypdf>=5.4.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.4.0",
]
